Added DB suffix generator to LocalDb
Upgraded LocalDb library to 0.2.0 to support datbase suffix feature
that will allow sandbox users to generate the suffix appeneded
to the database name.

The new default, the suffix is now {Guid}_{DateTime.Now.Ticks}. This
will garauntee uniqueness of database names, while still allowing users
to narrow down a database to a particular time (if necessary).

Bumped Sandbox version to 0.2.0